+++ NEWS.md
@@ -73,8 +73,8 @@
 
 ## Other new features
 
-* New `rig self update` to update rig, if it was installed via the install
-  script.
+* New `rig self update` to update rig, new `rig self uninstall` to remove
+  rig, if it was installed via the install script.
 
 * New `rig cache` command to inspect and manage rig's download and build
   caches. rig now reflinks or hardlinks packages from the cache when